requirements.txt
pyyaml
boto
